Add files via upload
Here I have Cat as a user in CN and John as an user in MN .The task was to install vsftpd software/package in MN through CN  by writing inventory and cfg file. We have disabled root ie do not use root for task instead give some powers to user Cat and John and then run the task.